particle-swarm-opt && GA

+++ particle_swarm/ParticleSwarmOptimizationProblem.java
@@ -10,7 +10,7 @@ public interface ParticleSwarmOptimizationProblem<T> {
 
     Set<T> initializePopulation(int size);
     boolean getStopCondition();
-    float eval(T candidate);
+    double eval(T candidate);
     void iterate(T solution, Integer timer);
     T randomVelocity(T particle);
     void updateVelocity(Particle<T> current, T bestSolution);
